// acheev-api/src/types/gqlTypings.generated.ts
export type InputMaybe<T> = T | undefined;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  Date: any;
  DateTime: any;
  JSON: any;
  Time: any;
};
// ...
export enum ChatEvents {
  DeleteMessage = 'DELETE_MESSAGE',
  NewMessage = 'NEW_MESSAGE',
  UpdateMessage = 'UPDATE_MESSAGE'
}
// ...
export type OrderBy = {
  column: Scalars['String'];
  direction?: InputMaybe<OrderByDirection>;
};

export enum OrderByDirection {
  Asc = 'ASC',
  Desc = 'DESC'
}